+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Yii 2 Application Runtime Configuration extension Change Log
 - Enh: Added `Manager::getItemValue()` allowing to get particular item value (klimov-paul)
 - Enh: Added `Manager::$autoRestoreValues` allowing to automatically restore item values from storage during manager initialization (klimov-paul)
 - Enh: `Manager::$cacheId` can now be specified as a PHP callback (klimov-paul)
+- Enh: Added `Item::$source` and `Manager::$source` allowing usage of the any object as config values source, not only current Yii application (klimov-paul)
 - Enh #9: `StorageFilterTrait` created providing ability to use single entity for multiple configuration storage at `StorageDb`, `StorageMongoDb` and `StorageActiveRecord` (klimov-paul)